chore: move currencyId to common
