+++ src/common.rs
@@ -1,7 +1,9 @@
 #[allow(unused_imports)]
 use crate::{client::MercadoPagoClient, payments::types::PaymentCreateOptions};
+use iso_currency::Currency;
 use reqwest::Response;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use serde_enum_str::{Deserialize_enum_str, Serialize_enum_str};
 use thiserror::Error;
 
 /// Used for solving json responses from Mercado Pago. If there is an error, [`MercadoPagoRequestError`] handles both the request body errors from Mercado Pago and Reqwest errors.
@@ -32,11 +34,73 @@ pub struct MercadoPagoError {
     /// Resume of the error
     pub message: String,
     /// Identification of the error. It usually has to do with the HTTP status.
-    pub error: String,
+    pub error: Option<String>,
     /// HTTP Status
     pub status: u16,
     /// A list of causes of the error
-    pub cause: Vec<MercadoPagoErrorCause>,
+    pub cause: Option<Vec<MercadoPagoErrorCause>>,
+}
+
+#[derive(Deserialize_enum_str, Serialize_enum_str, Debug, PartialEq, Eq, Default)]
+pub enum CurrencyId {
+    ARS,
+    #[default]
+    BRL,
+    CLP,
+    MXN,
+    COP,
+    PEN,
+    UYU,
+    VES,
+    MCN,
+    BTC,
+    USD,
+    USDP,
+    DCE,
+    ETH,
+    FDI,
+    CDB,
+    #[serde(other)]
+    Unknown(String),
+}
+
+impl From<Currency> for CurrencyId {
+    /// Parse ISO currency to `CurrencyId`
+    fn from(value: Currency) -> Self {
+        match value {
+            Currency::ARS => Self::ARS,
+            Currency::BRL => Self::BRL,
+            Currency::CLP => Self::CLP,
+            Currency::MXN => Self::MXN,
+            Currency::COP => Self::COP,
+            Currency::PEN => Self::PEN,
+            Currency::UYU => Self::UYU,
+            Currency::VES => Self::VES,
+            Currency::USD => Self::USD,
+            _ => Self::Unknown(value.to_string()),
+        }
+    }
+}
+
+impl TryFrom<CurrencyId> for Currency {
+    type Error = String;
+
+    /// Try parse `CurrencyId` to ISO currency
+    fn try_from(value: CurrencyId) -> Result<Self, Self::Error> {
+        match value {
+            CurrencyId::ARS => Ok(Currency::ARS),
+            CurrencyId::BRL => Ok(Currency::BRL),
+            CurrencyId::CLP => Ok(Currency::CLP),
+            CurrencyId::MXN => Ok(Currency::MXN),
+            CurrencyId::COP => Ok(Currency::COP),
+            CurrencyId::PEN => Ok(Currency::PEN),
+            CurrencyId::UYU => Ok(Currency::UYU),
+            CurrencyId::VES => Ok(Currency::VES),
+            CurrencyId::USD => Ok(Currency::USD),
+            CurrencyId::Unknown(v) => v.parse::<Currency>().map_err(|e| e.to_string()),
+            v => Err(format!("Unsupported currency: {v}")),
+        }
+    }
 }
 
 /// Error cause
